pyproject.toml
[project]
name = "honeyquant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-generativeai>=0.3.0",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
    "pyarrow>=14.0.0",
]
